# Faster proxs and omegas