This PR should be able to enable the following speed-up on the computation of omega and prox functions:
- about **450x** for `prox_group_sparsity` (from about 20.5 s to 48.2 ms)
- about **1400x** for `omega_group_sparsity` (from about 15.6 s to 11.2 ms)
The acceleration is achieved through a redesign of the data structure about the groups which allows a faster direct access to memory via `cython`.