@daducci 2018-11-16
@MarioOcampo , can you please check if this modification enable similar accelerations on your computer?