@MarioOcampo 2018-11-17
@daducci, the acceleration in my computer is:
- `prox_group_sparsity`: about **80x** (from about 16.6 s to 196.2 ms)
- `omega_group_sparsity`: about **400x** (from about 12.7 s to 32.6 ms)